requirements.txt
streamlit==1.55.0
groq==1.1.1
python-dotenv==1.2.2
